sidecar/pyproject.toml
[project]
name = "aletheia-sidecar"
version = "0.1.0"
description = "Local perceptual daemon for Aletheia"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymupdf>=1.23.0",
    "Pillow>=10.2.0",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.9.0",
    "layoutparser>=0.3.4",
    "camelot-py[cv]>=0.11.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
ml = [
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "easyocr>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
